fixed the bug with Resources not showing when changing language
