+++ lib/blocs/course_list/course_list_bloc.dart
@@ -8,7 +8,6 @@ import 'package:bro/data/course_repository.dart';
 import 'package:bro/data/preferred_language_repository.dart';
 import 'package:bro/models/courses.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:graphql_flutter/graphql_flutter.dart';
 
 /// CourseListBloc is in charge of pulling courses from the backend.
 /// It listens to PreferredLanguage State and will refresh the courses when the state is changed.
@@ -70,27 +69,17 @@ class CourseListBloc extends Bloc<CourseListEvent, CourseListState> {
   /// Returns all courses in the preferredLanguage appended with courses in Norwegian.
   Future<CourseListState> _retrieveCourses(
       CourseListEvent event, int curr_len) async {
-    try {
-      var langSlug;
-      event is CourseListRefresh
-          ? langSlug = event.preferredLang
-          : langSlug = await preferredLanguageRepository.getPreferredLangSlug();
-      return await repository
-          .getLangCourses(langSlug, curr_len, 10, recommended)
-          .then((res) {
-        var res_list = List<Map<String, dynamic>>.from(res.data!['LangCourse']);
-        final returnCourse = LangCourseList.takeList(res_list).langCourses;
-        return Success(courses: returnCourse, hasReachedMax: false);
-      });
-    } on NetworkException catch (e, stackTrace) {
-      log(e.toString());
-      log(stackTrace.toString());
-      return Failed();
-    } catch (e, stackTrace) {
-      log(e.toString());
-      log(stackTrace.toString());
-      return Failed();
-    }
+    var langSlug;
+    event is CourseListRefresh
+        ? langSlug = event.preferredLang
+        : langSlug = await preferredLanguageRepository.getPreferredLangSlug();
+    return await repository
+        .getLangCourses(langSlug, curr_len, 10, recommended)
+        .then((res) {
+      var res_list = List<Map<String, dynamic>>.from(res.data!['LangCourse']);
+      final returnCourse = LangCourseList.takeList(res_list).langCourses;
+      return Success(courses: returnCourse, hasReachedMax: false);
+    });
   }
 }
