+++ lib/blocs/home/home_bloc.dart
@@ -1,14 +1,11 @@
 import 'dart:async';
-import 'dart:developer';
 import 'package:bro/blocs/home/home_bucket.dart';
 import 'package:bro/blocs/home/home_state.dart';
 import 'package:bro/data/home_repository.dart';
 import 'package:bro/models/home.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:graphql_flutter/graphql_flutter.dart';
 
-/// Homerequested contains home, recommended courses and recommended resources.
-/// Listens to changes in PreferredLanguageState and updates based on changes.
+/// Homerequested contains home
 class HomeBloc extends Bloc<HomeEvent, HomeState> {
   HomeRepository homeRepository;
 
@@ -16,7 +13,6 @@ class HomeBloc extends Bloc<HomeEvent, HomeState> {
 
   @override
   Stream<HomeState> mapEventToState(HomeEvent event) async* {
-    // Not able to access state methods without this. Do not know why.
     if (event is HomeRequested) {
       yield await _retrieveCourses(event, 0);
     }
@@ -28,13 +24,7 @@ class HomeBloc extends Bloc<HomeEvent, HomeState> {
       var home = await homeRepository.getHome();
       var returnHome = Home.fromJson(home.data!['home']);
       return HomeSuccess(home: returnHome);
-    } on NetworkException catch (e, stackTrace) {
-      log(e.toString());
-      log(stackTrace.toString());
-      return Failed();
-    } catch (e, stackTrace) {
-      log(e.toString());
-      log(stackTrace.toString());
+    } catch (e) {
       return Failed();
     }
   }
