+++ lib/data/course_repository.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'package:bro/data/queries/home_view_queries.dart';
 import 'package:bro/data/queries/queries.dart';
 import 'package:gql/language.dart';
 import 'package:graphql_flutter/graphql_flutter.dart';
