+++ lib/data/home_repository.dart
@@ -9,41 +9,10 @@ class HomeRepository {
 
   HomeRepository({required this.client});
 
-  // Course type should be made in a models/ directory
-  Future<QueryResult> getRecommendedCourses(
-      String lang_slug, int start, int limit) async {
-    final _options = WatchQueryOptions(
-      document: parseString(getRecommendedCoursesQuery),
-      fetchResults: true,
-      variables: <String, dynamic>{
-        'lang_slug': lang_slug,
-        'start': start,
-        'limit': limit
-      },
-    );
-
-    return await client.query(_options);
-  }
-
   Future<QueryResult> getHome() async {
     final _options = WatchQueryOptions(
       document: parseString(getHomeQuery),
       fetchResults: true,
-    );
-
-    return await client.query(_options);
-  }
-
-  Future<QueryResult> getRecommendedLangResources(
-      int start, int limit, String lang) async {
-    final _options = WatchQueryOptions(
-      document: parseString(getRecommendedLangResourcesQuery),
-      fetchResults: true,
-      variables: <String, dynamic>{
-        'start': start,
-        'limit': limit,
-        'lang': lang
-      },
     );
 
     return await client.query(_options);
