+++ lib/data/queries/resource_queries.dart
@@ -92,3 +92,67 @@ query($lang: String, $category: Int) {
     }
   }
 }''';
+
+final String getResourcesQuery = r'''
+query($lang: String, $category: Int) {
+  LangResource: resources(
+    where: {
+      _where: [
+        { language: { slug: $lang } }
+        { category: { id: $category } }
+      ]
+    }
+  ) {
+    title
+    cover_photo {url}
+    description
+    language {slug}
+    resource_group {slug}
+    description
+    publisher {name}
+    category {category_name id}
+    is_recommended
+    references {
+      reference_title
+      reference_description
+      reference_url
+      reference_button_text
+    }
+    documents {
+      document_name 
+      document_file {url}
+    }
+  }
+}''';
+
+final String getRecommendedLangResourcesQuery = r'''
+query recommendedLangResourcesQuery ($start: Int!, $limit: Int!,$lang: String) {
+  LangResource: resources(start: $start, limit: $limit,
+    where: {
+      _where: [
+        { language: { slug: $lang } }
+        {is_recommended:true} 
+      ]
+    }
+  ) {
+    title
+    cover_photo {url}
+    description
+    language {slug}
+    resource_group {slug}
+    description
+    publisher {name}
+    category {category_name id}
+    is_recommended
+    references {
+      reference_title
+      reference_description
+      reference_url
+      reference_button_text
+    }
+    documents {
+      document_name 
+      document_file {url}
+    }
+  }
+}''';
