+++ lib/data/resource_repository.dart
@@ -1,4 +1,4 @@
-import 'package:bro/data/queries/home_view_queries.dart';
+
 import 'package:bro/data/queries/resource_queries.dart';
 import 'package:gql/language.dart';
 import 'package:graphql_flutter/graphql_flutter.dart';
@@ -19,6 +19,32 @@ class ResourceRepository {
   }
 
   Future<QueryResult> getLangResources(
+      String lang_slug, String category,bool recommended) async {
+    final _options;
+    recommended ? _options = WatchQueryOptions(
+      document: parseString(getRecommendedLangResourcesQuery),
+      fetchResults: true,
+      variables: <String, dynamic>{
+        'lang': lang_slug
+        'start': 0,
+        'limit': 3
+      },
+    ) :
+    _options = WatchQueryOptions(
+      document: parseString(getResourcesQuery),
+      fetchResults: true,
+      variables: <String, dynamic>{
+        'lang': lang_slug,
+        'category': int.parse(category),
+/*         'start': start,
+        'limit': limit */
+      },
+    );
+
+    return await client.query(_options);
+  }
+
+  Future<QueryResult> getFalseLangResources(
       String lang_slug, String category,bool recommended) async {
     final _options;
     recommended ? _options = WatchQueryOptions(
